third_party/blink/web_tests/fast/forms/interactive-validation-prevented-expected.txt

CONSOLE ERROR: An invalid form control with name='i1' is not focusable.
Test if an invalid control prevents interactive form submission, and the first invalid focusable control gets focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS The form was not submitted
PASS document.activeElement is document.getElementById("i2")
TEST COMPLETE
  
